+++ cupy/__init__.py
@@ -54,7 +54,6 @@ __version__ = _version.__version__
 
 from cupy import binary  # NOQA
 import cupy.core.fusion  # NOQA
-from cupy import creation  # NOQA
 from cupy import fft  # NOQA
 from cupy import functional  # NOQA
 from cupy import indexing  # NOQA
@@ -252,36 +251,36 @@ from builtins import complex  # NOQA
 # -----------------------------------------------------------------------------
 # Array creation routines
 # -----------------------------------------------------------------------------
-from cupy.creation.basic import empty  # NOQA
-from cupy.creation.basic import empty_like  # NOQA
-from cupy.creation.basic import eye  # NOQA
-from cupy.creation.basic import full  # NOQA
-from cupy.creation.basic import full_like  # NOQA
-from cupy.creation.basic import identity  # NOQA
-from cupy.creation.basic import ones  # NOQA
-from cupy.creation.basic import ones_like  # NOQA
-from cupy.creation.basic import zeros  # NOQA
-from cupy.creation.basic import zeros_like  # NOQA
+from cupy._creation.basic import empty  # NOQA
+from cupy._creation.basic import empty_like  # NOQA
+from cupy._creation.basic import eye  # NOQA
+from cupy._creation.basic import full  # NOQA
+from cupy._creation.basic import full_like  # NOQA
+from cupy._creation.basic import identity  # NOQA
+from cupy._creation.basic import ones  # NOQA
+from cupy._creation.basic import ones_like  # NOQA
+from cupy._creation.basic import zeros  # NOQA
+from cupy._creation.basic import zeros_like  # NOQA
 
-from cupy.creation.from_data import copy  # NOQA
-from cupy.creation.from_data import array  # NOQA
-from cupy.creation.from_data import asanyarray  # NOQA
-from cupy.creation.from_data import asarray  # NOQA
-from cupy.creation.from_data import ascontiguousarray  # NOQA
-from cupy.creation.from_data import fromfile  # NOQA
+from cupy._creation.from_data import copy  # NOQA
+from cupy._creation.from_data import array  # NOQA
+from cupy._creation.from_data import asanyarray  # NOQA
+from cupy._creation.from_data import asarray  # NOQA
+from cupy._creation.from_data import ascontiguousarray  # NOQA
+from cupy._creation.from_data import fromfile  # NOQA
 
-from cupy.creation.ranges import arange  # NOQA
-from cupy.creation.ranges import linspace  # NOQA
-from cupy.creation.ranges import logspace  # NOQA
-from cupy.creation.ranges import meshgrid  # NOQA
-from cupy.creation.ranges import mgrid  # NOQA
-from cupy.creation.ranges import ogrid  # NOQA
+from cupy._creation.ranges import arange  # NOQA
+from cupy._creation.ranges import linspace  # NOQA
+from cupy._creation.ranges import logspace  # NOQA
+from cupy._creation.ranges import meshgrid  # NOQA
+from cupy._creation.ranges import mgrid  # NOQA
+from cupy._creation.ranges import ogrid  # NOQA
 
-from cupy.creation.matrix import diag  # NOQA
-from cupy.creation.matrix import diagflat  # NOQA
-from cupy.creation.matrix import tri  # NOQA
-from cupy.creation.matrix import tril  # NOQA
-from cupy.creation.matrix import triu  # NOQA
+from cupy._creation.matrix import diag  # NOQA
+from cupy._creation.matrix import diagflat  # NOQA
+from cupy._creation.matrix import tri  # NOQA
+from cupy._creation.matrix import tril  # NOQA
+from cupy._creation.matrix import triu  # NOQA
 
 # -----------------------------------------------------------------------------
 # Functional routines
