+++ cupy/_creation/from_data.py
