+++ cupy/_creation/matrix.py
