+++ cupy/_creation/ranges.py
