--- cupy/creation/__init__.py
@@ -1,8 +1,0 @@
-# Functions from the following NumPy document
-# https://docs.scipy.org/doc/numpy/reference/routines.array-creation.html
-
-# "NOQA" to suppress flake8 warning
-from cupy.creation import basic  # NOQA
-from cupy.creation import from_data  # NOQA
-from cupy.creation import matrix  # NOQA
-from cupy.creation import ranges  # NOQA
